requirements.txt
python-telegram-bot>=20.0
aiohttp
peewee
beautifulsoup4
nest_asyncio
